Update the main files and src folder
